% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2r-package.R
\docType{package}
\name{api2r-package}
\alias{api2r}
\alias{api2r-package}
\title{api2r: Rapidly Scaffold API Client Packages}
\description{
Automatically generate R package skeletons from 'application programming interfaces (APIs)' that follow the 'OpenAPI Specification (OAS)'. The skeletons implement best practices to streamline package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/api2r}
  \item \url{https://jonthegeek.github.io/api2r/}
  \item Report bugs at \url{https://github.com/jonthegeek/api2r/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
