% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bk_data.R
\name{as_bk_data}
\alias{as_bk_data}
\title{Prepare rapid objects for beekeeper}
\usage{
as_bk_data(x, ...)
}
\arguments{
\item{x}{(\code{any}) The object to check or coerce.}

\item{...}{Additional parameters passed to methods.}
}
\value{
(\code{list}) A list.
}
\description{
Convert \code{rapid} objects to lists of properties to use in beekeeper templates.
}
\keyword{internal}
