% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beekeeper-package.R
\docType{package}
\name{beekeeper-package}
\alias{beekeeper}
\alias{beekeeper-package}
\title{beekeeper: Rapidly Scaffold API Client Packages}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Automatically generate R package skeletons from 'application programming interfaces (APIs)' that follow the 'OpenAPI Specification (OAS)'. The skeletons implement best practices to streamline package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beekeeper.api2r.org}
  \item \url{https://github.com/api2r/beekeeper}
  \item Report bugs at \url{https://github.com/api2r/beekeeper/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
\keyword{internal}
