% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beekeeper.R
\name{.assert_config_exists}
\alias{.assert_config_exists}
\title{Assert the beekeeper config file exists}
\usage{
.assert_config_exists(pkg_dir, config_filename, call = caller_env())
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{call}{(\code{environment}) The caller environment for error messages.}
}
\value{
(\code{NULL}, invisibly) Called for side effect of asserting the config
file exists.
}
\description{
Assert the beekeeper config file exists
}
\keyword{internal}
