% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-template.R
\name{.bk_use_template}
\alias{.bk_use_template}
\title{Use a template in this package}
\usage{
.bk_use_template(
  template,
  data,
  ...,
  target = template,
  dir = c("R", "tests/testthat")
)
}
\arguments{
\item{template}{(\code{character(1)}) The template file name to render.}

\item{data}{(\code{list}) Data passed to a template.}

\item{...}{These dots are for future extensions and must be empty.}

\item{target}{(\code{character(1)}) The name of the generated file.}

\item{dir}{(\code{character(1)}) The directory where a generated file should be
written.}
}
\value{
The path to the generated or updated file, invisibly.
}
\description{
Use a template in this package
}
\keyword{internal}
