% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-template.R
\name{.bk_use_template_impl}
\alias{.bk_use_template_impl}
\title{Write a rendered template to disk}
\usage{
.bk_use_template_impl(template, data, target, dir)
}
\arguments{
\item{template}{(\code{character(1)}) The template file name to render.}

\item{data}{(\code{list}) Data passed to a template.}

\item{target}{(\code{character(1)}) The name of the generated file.}

\item{dir}{(\code{character(1)}) The directory where a generated file should be
written.}
}
\value{
(\code{character(1)}) The generated file path.
}
\description{
Write a rendered template to disk
}
\keyword{internal}
