% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-api_abbr.R
\name{.check_api_abbr}
\alias{.check_api_abbr}
\title{Check whether an object is an API abbreviation}
\usage{
.check_api_abbr(api_abbr)
}
\arguments{
\item{api_abbr}{An abbreviation to use for this API, such as \code{"goog"} or
\code{"asn"}.}
}
\value{
The input, if it is valid.
}
\description{
Check whether an object is an API abbreviation
}
\keyword{internal}
