% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-000.R
\name{.check_character}
\alias{.check_character}
\title{Check whether an object is a character vector}
\usage{
.check_character(
  x,
  expected = "character",
  arg = rlang::caller_arg(url),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{The object to check.}

\item{expected}{The expected argument class or type.}

\item{arg}{The name of the argument.}

\item{call}{The environment from which the checker was called.}
}
\value{
The input, if it is valid.
}
\description{
Check whether an object is a character vector
}
\keyword{internal}
