% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-000.R
\name{.check_error}
\alias{.check_error}
\title{Generate checker error messages}
\usage{
.check_error(
  expected,
  arg = rlang::caller_arg(url),
  call = rlang::caller_env()
)
}
\arguments{
\item{expected}{The expected argument class or type.}

\item{arg}{The name of the argument.}

\item{call}{The environment from which the checker was called.}
}
\description{
Generate checker error messages
}
\keyword{internal}
