% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-000.R
\name{.check_scalar}
\alias{.check_scalar}
\title{Check whether an object has length 1}
\usage{
.check_scalar(
  x,
  expected,
  arg = rlang::caller_arg(url),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{The object to check.}

\item{expected}{The expected argument class or type.}

\item{arg}{The name of the argument.}

\item{call}{The environment from which the checker was called.}
}
\value{
The input, if it is valid.
}
\description{
Check whether an object has length 1
}
\keyword{internal}
