% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args-url.R
\name{.check_url_scalar}
\alias{.check_url_scalar}
\title{Check whether an object is a URL}
\usage{
.check_url_scalar(
  url,
  arg = rlang::caller_arg(url),
  call = rlang::caller_env()
)
}
\arguments{
\item{url}{A url, such as \code{"https://www.example.com"}.}

\item{arg}{The name of the argument.}

\item{call}{The environment from which the checker was called.}
}
\value{
The input, if it is valid.
}
\description{
Check whether an object is a URL
}
\keyword{internal}
