% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.collapse_comma}
\alias{.collapse_comma}
\title{Collapse to a comma-separated string}
\usage{
.collapse_comma(to_collapse)
}
\arguments{
\item{to_collapse}{(\code{character}) The character vector to collapse.}
}
\value{
(\code{glue}) A length-1, comma-separated glue object.
}
\description{
Collapse to a comma-separated string
}
\keyword{internal}
