% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.collapse_comma_newline}
\alias{.collapse_comma_newline}
\title{Collapse to a comma-separated vertical string}
\usage{
.collapse_comma_newline(to_collapse)
}
\arguments{
\item{to_collapse}{(\code{character}) The character vector to collapse.}
}
\value{
A length-1, comma-separated glue object with newlines.
}
\description{
Collapse to a comma-separated vertical string
}
\keyword{internal}
