% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.collapse_comma_self_equal}
\alias{.collapse_comma_self_equal}
\title{Collapse to a comma-separated x = x string}
\usage{
.collapse_comma_self_equal(x)
}
\value{
A length-1, comma-separated glue object.
}
\description{
Collapse to a comma-separated x = x string
}
\keyword{internal}
