% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.collapse_quote_comma}
\alias{.collapse_quote_comma}
\title{Collapse to a comma-separated quoted string}
\usage{
.collapse_quote_comma(to_collapse)
}
\arguments{
\item{to_collapse}{(\code{character}) The character vector to collapse.}
}
\value{
A length-1, comma-separated glue object.
}
\description{
Collapse to a comma-separated quoted string
}
\keyword{internal}
