% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.compile_param_class_descriptions}
\alias{.compile_param_class_descriptions}
\title{Compile parameter class descriptions}
\usage{
.compile_param_class_descriptions(type, allow_empty, required)
}
\arguments{
\item{type}{(\code{data.frame}) Joined parameter type metadata.}

\item{allow_empty}{(\code{logical}) Whether empty parameter values may be
represented as \code{NULL}.}

\item{required}{(\code{logical}) Whether each parameter is required.}
}
\value{
(\code{character}) Parameter class descriptions.
}
\description{
Compile parameter class descriptions
}
\keyword{internal}
