% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beekeeper.R
\name{.config_origin}
\alias{.config_origin}
\title{Extract the origin from an api definition}
\usage{
.config_origin(api_definition)
}
\arguments{
\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}
}
\value{
(\code{list}) The \code{url}, \code{format}, and \code{version} (if known) of the API
definition.
}
\description{
Extract the origin from an api definition
}
\keyword{internal}
