% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.describe_param_classes}
\alias{.describe_param_classes}
\title{Describe parameter classes}
\usage{
.describe_param_classes(params_schema, allow_empty, required)
}
\arguments{
\item{params_schema}{(\code{data.frame}) Schema metadata for operation parameters.}

\item{allow_empty}{(\code{logical}) Whether empty parameter values may be
represented as \code{NULL}.}

\item{required}{(\code{logical}) Whether each parameter is required.}
}
\value{
(\code{character}) Display strings for parameter classes.
}
\description{
Describe parameter classes
}
\keyword{internal}
