% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.extract_params_by_location}
\alias{.extract_params_by_location}
\title{Extract parameter names by location}
\usage{
.extract_params_by_location(params_df, filter_in)
}
\arguments{
\item{params_df}{(\code{data.frame}) Parameter metadata in tabular form.}

\item{filter_in}{(\code{character(1)}) The parameter location to keep.}
}
\value{
(\code{character}) Parameter names for the requested location.
}
\description{
Extract parameter names by location
}
\keyword{internal}
