% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oas.R
\name{.fetch_oas}
\alias{.fetch_oas}
\title{Fetch an OpenAPI Spec}
\usage{
.fetch_oas(oas_url)
}
\arguments{
\item{oas_url}{A single character giving the URL of the spec.}
}
\value{
An object with class \code{oas}, representing the API spec.
}
\description{
Fetch an OpenAPI Spec
}
\keyword{internal}
