% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.flatten_df_data_frame}
\alias{.flatten_df_data_frame}
\title{Flatten a data frame method}
\usage{
.flatten_df_data_frame(x)
}
\arguments{
\item{x}{(\code{data.frame}) The object to return unchanged.}
}
\value{
A \code{data.frame}.
}
\description{
Flatten a data frame method
}
\keyword{internal}
