% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.flatten_df_list}
\alias{.flatten_df_list}
\title{Flatten a list method}
\usage{
.flatten_df_list(x)
}
\arguments{
\item{x}{(\code{list}) The objects to row-bind.}
}
\value{
A \code{data.frame}.
}
\description{
Flatten a list method
}
\keyword{internal}
