% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.flatten_df_null}
\alias{.flatten_df_null}
\title{Flatten a \code{NULL} method}
\usage{
.flatten_df_null(x)
}
\arguments{
\item{x}{(\code{NULL}) The object to convert.}
}
\value{
A \code{data.frame}.
}
\description{
Flatten a \code{NULL} method
}
\keyword{internal}
