% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.format_subset_path_str}
\alias{.format_subset_path_str}
\title{Format a subset_path vector as an R string literal}
\usage{
.format_subset_path_str(subset_path)
}
\value{
(\code{character(1)}) Either \code{'"name"'} or \code{'c("a", "b")'}.
}
\description{
Format a subset_path vector as an R string literal
}
\keyword{internal}
