% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.generate_paths_files}
\alias{.generate_paths_files}
\title{Generate files for all operations}
\usage{
.generate_paths_files(
  paths_by_operation,
  api_abbr,
  security_data,
  pagination_data
)
}
\arguments{
\item{paths_by_operation}{(\code{list}) Template-ready operations keyed by
operation identifier.}

\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{security_data}{(\code{list}) Generated security metadata.}

\item{pagination_data}{(\code{list}) Pagination metadata used while generating
path files.}
}
\value{
(\code{character}) Generated file paths.
}
\description{
Generate files for all operations
}
\keyword{internal}
