% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg.R
\name{.generate_pkg_impl}
\alias{.generate_pkg_impl}
\title{Generate package files from prepared inputs}
\usage{
.generate_pkg_impl(config, api_definition, security_data)
}
\arguments{
\item{config}{(\code{list}) Package-generation configuration data.}

\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{security_data}{(\code{list}) Generated security metadata.}
}
\value{
(\code{character}, invisibly) Generated file paths.
}
\description{
Generate package files from prepared inputs
}
\keyword{internal}
