% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_req_prepare.R
\name{.generate_prepare}
\alias{.generate_prepare}
\title{Generate prepare files}
\usage{
.generate_prepare(config, api_definition, security_data, pkg_dir = ".")
}
\arguments{
\item{config}{(\code{list}) Package-generation configuration data.}

\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{security_data}{(\code{list}) Generated security metadata.}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character}) Generated file paths.
}
\description{
Generate prepare files
}
\keyword{internal}
