% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-prepare.R
\name{.generate_prepare_r}
\alias{.generate_prepare_r}
\title{Generate the prepare helper}
\usage{
.generate_prepare_r(config, api_definition, security_data)
}
\arguments{
\item{config}{(\code{list}) Package-generation configuration data.}

\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{security_data}{(\code{list}) Generated security metadata.}
}
\value{
(\code{character(1)}) The generated file path.
}
\description{
Generate the prepare helper
}
\keyword{internal}
