% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-security.R
\name{.generate_security}
\alias{.generate_security}
\title{Generate security files and metadata}
\usage{
.generate_security(api_abbr, security_schemes)
}
\arguments{
\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{security_schemes}{(\code{rapid::class_security_schemes}) Security schemes
from the API definition.}
}
\value{
(\code{list}) Generated security metadata.
}
\description{
Generate security files and metadata
}
\keyword{internal}
