% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{.generate_security_signature}
\alias{.generate_security_signature}
\title{Generate a security argument signature}
\usage{
.generate_security_signature(security_arg_names, api_abbr)
}
\arguments{
\item{security_arg_names}{(\code{character}) Security argument names to exclude
from generated operation signatures (because they are defined once,
separately).}

\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}
}
\value{
(\code{character(1)}) A function-signature fragment.
}
\description{
Generate a security argument signature
}
\keyword{internal}
