% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_shared_params.R
\name{.generate_shared_params}
\alias{.generate_shared_params}
\title{Generate shared parameter docs}
\usage{
.generate_shared_params(security_data, pkg_dir = ".")
}
\arguments{
\item{security_data}{(\code{list}) Generated security metadata.}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}) The generated file path.
}
\description{
Generate shared parameter docs
}
\keyword{internal}
