% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.glue_pipe_brace}
\alias{.glue_pipe_brace}
\title{Interpolate glue with pipe braces}
\usage{
.glue_pipe_brace(..., .envir = rlang::caller_env())
}
\arguments{
\item{...}{Expressions passed to \code{\link[glue:glue]{glue::glue()}}.}

\item{.envir}{(\code{environment}) The environment in which to evaluate \code{...}.}
}
\value{
(\code{glue}) A glue object.
}
\description{
Interpolate glue with pipe braces
}
\keyword{internal}
