% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.is_pkg}
\alias{.is_pkg}
\title{Check whether we're in a package}
\usage{
.is_pkg(pkg_dir = usethis::proj_get())
}
\value{
(\code{logical(1)}) \code{TRUE} if the project is a package, \code{FALSE} if not.
}
\description{
Inspired by usethis:::is_package.
}
\keyword{internal}
