% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.maybe_use_stbl}
\alias{.maybe_use_stbl}
\title{Add stbl to dependencies if needed}
\usage{
.maybe_use_stbl(pkg_dir, paths, security_arg_names)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{paths}{(\code{rapid::class_paths}) API path definitions.}

\item{security_arg_names}{(\code{character}) Security argument names to exclude
from generated operation signatures (because they are defined once,
separately).}
}
\value{
(\code{character(1)} or \code{NULL}, invisibly) "stbl" if stbl is used, \code{NULL}
otherwise.
}
\description{
Add stbl to dependencies if needed
}
\keyword{internal}
