% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.maybe_use_tibblify}
\alias{.maybe_use_tibblify}
\title{Add tibblify to dependencies if needed}
\usage{
.maybe_use_tibblify(pkg_dir, paths)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{paths}{(\code{rapid::class_paths}) API path definitions.}
}
\value{
(\code{character(1)} or \code{NULL}, invisibly) "tibblify" if tibblify is
used, \code{NULL} otherwise.
}
\description{
Add tibblify to dependencies if needed
}
\keyword{internal}
