% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oas.R
\name{.new_oas}
\alias{.new_oas}
\title{Create an oas object}
\usage{
.new_oas(x = list())
}
\arguments{
\item{x}{A list created by reading an OAS spec with read_yaml.}
}
\value{
That list with additional class \code{oas}.
}
\description{
Create an oas object
}
\keyword{internal}
