% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Register package methods on load}
\usage{
.onLoad(...)
}
\arguments{
\item{...}{Additional arguments passed by R during package loading.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Register package methods on load
}
\keyword{internal}
