% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beekeeper.R
\name{.optional_origin_field}
\alias{.optional_origin_field}
\title{Get an origin field if it exists}
\usage{
.optional_origin_field(api_definition, field)
}
\arguments{
\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{field}{(\code{character}) The name of the field to extract.}
}
\value{
(\code{any}, most likely \code{character(1)} or \code{NULL}) The value of the
field, if it exists in the \code{origin} property of the \code{info} property of
\code{api_definition} object.
}
\description{
Get an origin field if it exists
}
\keyword{internal}
