% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beekeeper.R
\name{.optional_S7_prop}
\alias{.optional_S7_prop}
\title{Get an S7 property if it exists}
\usage{
.optional_S7_prop(obj, prop)
}
\arguments{
\item{obj}{(\code{character}) The S7 object from which the property should be
extracted.}

\item{prop}{(\code{character}) The name of the property to extract.}
}
\value{
(\code{any}) The value of the property, if it exists in \code{obj} (\code{NULL}
invisibly otherwise).
}
\description{
Get an S7 property if it exists
}
\keyword{internal}
