% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.param_schema_to_r}
\alias{.param_schema_to_r}
\title{Map parameter schemas to \code{stbl} helpers}
\usage{
.param_schema_to_r(params_schema)
}
\arguments{
\item{params_schema}{(\code{data.frame}) Schema metadata for operation
parameters.}
}
\value{
(\code{character}) Coercion helper names.
}
\description{
Map parameter schemas to \code{stbl} helpers
}
\keyword{internal}
