% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.params_to_args}
\alias{.params_to_args}
\title{Convert parameters to a signature string}
\usage{
.params_to_args(params)
}
\arguments{
\item{params}{(\code{list}) Parameter metadata for a single operation.}
}
\value{
(\code{character(1)}) A comma-separated parameter string.
}
\description{
Convert parameters to a signature string
}
\keyword{internal}
