% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.params_to_list}
\alias{.params_to_list}
\title{Convert parameter rows to a list}
\usage{
.params_to_list(params_df)
}
\arguments{
\item{params_df}{(\code{data.frame}) Parameter metadata in tabular form.}
}
\value{
(\code{list}) Parameter metadata entries.
}
\description{
Convert parameter rows to a list
}
\keyword{internal}
