% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.params_to_named_args}
\alias{.params_to_named_args}
\title{Convert parameters to named arguments}
\usage{
.params_to_named_args(params)
}
\arguments{
\item{params}{(\code{list}) Parameter metadata for a single operation.}
}
\value{
(\code{character(1)}) A comma-separated named-argument string.
}
\description{
Convert parameters to named arguments
}
\keyword{internal}
