% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.params_to_validations}
\alias{.params_to_validations}
\title{Convert parameters to validation metadata}
\usage{
.params_to_validations(params)
}
\arguments{
\item{params}{(\code{list}) Parameter metadata for a single operation.}
}
\value{
(\code{list}) Validation metadata.
}
\description{
Convert parameters to validation metadata
}
\keyword{internal}
