% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.paste0_if}
\alias{.paste0_if}
\title{Append text conditionally}
\usage{
.paste0_if(original, test, addition)
}
\arguments{
\item{original}{(\code{character}) The original text.}

\item{test}{(\code{logical}) A condition for each element of \code{original}.}

\item{addition}{(\code{character}) Text to append when \code{test} is \code{TRUE}.}
}
\value{
(\code{character}) The updated text.
}
\description{
Append text conditionally
}
\keyword{internal}
