% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.path_as_arg}
\alias{.path_as_arg}
\title{Convert a path to a request argument}
\usage{
.path_as_arg(path, params_df)
}
\arguments{
\item{path}{(\code{character(1)}) The path template.}

\item{params_df}{(\code{data.frame}) Parameter metadata in tabular form.}
}
\value{
(\code{character(1)}) Code for the request path argument.
}
\description{
Convert a path to a request argument
}
\keyword{internal}
