% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.paths_fill_descriptions}
\alias{.paths_fill_descriptions}
\title{Fill missing operation descriptions}
\usage{
.paths_fill_descriptions(operation_descriptions, operation_summaries)
}
\arguments{
\item{operation_descriptions}{(\code{character}) Operation descriptions from the
API definition.}

\item{operation_summaries}{(\code{character}) Operation summaries from the API
definition.}
}
\value{
(\code{character}) Operation descriptions.
}
\description{
Fill missing operation descriptions
}
\keyword{internal}
