% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.paths_fill_operation_id}
\alias{.paths_fill_operation_id}
\title{Fill missing operation identifiers}
\usage{
.paths_fill_operation_id(operation_ids, endpoints, methods)
}
\arguments{
\item{operation_ids}{(\code{character}) Operation identifiers from the API
definition.}

\item{endpoints}{(\code{character}) Endpoint paths paired with operations.}

\item{methods}{(\code{character}) HTTP methods paired with endpoints.}
}
\value{
(\code{character}) Operation identifiers.
}
\description{
Fill missing operation identifiers
}
\keyword{internal}
