% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.paths_fill_summary}
\alias{.paths_fill_summary}
\title{Fill missing operation summaries}
\usage{
.paths_fill_summary(operation_summaries, endpoints, methods)
}
\arguments{
\item{operation_summaries}{(\code{character}) Operation summaries from the API
definition.}

\item{endpoints}{(\code{character}) Endpoint paths paired with operations.}

\item{methods}{(\code{character}) HTTP methods paired with endpoints.}
}
\value{
(\code{character}) Human-readable operation summaries.
}
\description{
Fill missing operation summaries
}
\keyword{internal}
