% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.paths_fill_tags}
\alias{.paths_fill_tags}
\title{Fill missing operation tags}
\usage{
.paths_fill_tags(tags)
}
\arguments{
\item{tags}{(\code{list}) Operation tags from the API definition.}
}
\value{
(\code{character}) Snake-case tag names.
}
\description{
Fill missing operation tags
}
\keyword{internal}
