% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.paths_need_tibblify}
\alias{.paths_need_tibblify}
\title{Determine whether generated paths need \code{tibblify}}
\usage{
.paths_need_tibblify(paths)
}
\arguments{
\item{paths}{(\code{rapid::class_paths}) API path definitions.}
}
\value{
(\code{logical(1)}) \code{TRUE} if any path has a JSON response spec.
}
\description{
Determine whether generated paths need \code{tibblify}
}
\keyword{internal}
