% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.paths_to_clean_df}
\alias{.paths_to_clean_df}
\title{Convert API paths to an operations data frame}
\usage{
.paths_to_clean_df(paths)
}
\arguments{
\item{paths}{(\code{rapid::class_paths}) API path definitions.}
}
\value{
(\code{tibble}) A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per non-deprecated
operation.
}
\description{
Convert API paths to an operations data frame
}
\keyword{internal}
