% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-conditions.R
\name{.pkg_abort}
\alias{.pkg_abort}
\title{Signal a package-scoped error}
\usage{
.pkg_abort(
  message,
  subclass,
  parent = NULL,
  call = rlang::caller_env(),
  message_env = rlang::caller_env(),
  ...
)
}
\arguments{
\item{message}{(\code{character}) The message for the new error. Messages will be
formatted with \code{\link[cli:cli_bullets]{cli::cli_bullets()}}.}

\item{subclass}{(\code{character}) Class(es) to assign to the error. Will be
prefixed by "\{package\}-error-".}

\item{parent}{A parent condition, as you might create during a
\code{\link[rlang:try_fetch]{rlang::try_fetch()}}. See \code{\link[rlang:abort]{rlang::abort()}} for additional information.}

\item{call}{(\code{environment}) The caller environment for error messages.}

\item{message_env}{(\code{environment}) The execution environment to use to
evaluate variables in error messages.}

\item{...}{Additional parameters passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} and on to
\code{\link[rlang:abort]{rlang::abort()}}.}
}
\value{
(\code{NULL}) Does not return.
}
\description{
Signal a package-scoped error
}
\keyword{internal}
