% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.prep_param_args}
\alias{.prep_param_args}
\title{Prepare parameter arguments}
\usage{
.prep_param_args(params, security_args)
}
\arguments{
\item{params}{(\code{list}) Parameter metadata for a single operation.}

\item{security_args}{(\code{character}) Security argument names to remove from
operation parameters (because they are defined once, separately).}
}
\value{
(\code{character(1)}) A comma-separated named-argument string.
}
\description{
Prepare parameter arguments
}
\keyword{internal}
