% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.prepare_params_df}
\alias{.prepare_params_df}
\title{Prepare parameter metadata}
\usage{
.prepare_params_df(params_df)
}
\arguments{
\item{params_df}{(\code{data.frame}) Parameter metadata in tabular form.}
}
\value{
(\code{data.frame}) Prepared parameter metadata.
}
\description{
Prepare parameter metadata
}
\keyword{internal}
