% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{.read_config_field}
\alias{.read_config_field}
\title{Read a field from beekeeper config}
\usage{
.read_config_field(field, pkg_dir, config_filename, default = NULL)
}
\arguments{
\item{field}{(\code{character(1)}) The name of the config field to read.}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{default}{(\code{any}) Default value to return if \code{field} is not in the
config.}
}
\value{
(\code{any}) The value of the requested field, or \code{default} if not found.
}
\description{
Read a field from beekeeper config
}
\keyword{internal}
