% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{.read_security_data}
\alias{.read_security_data}
\title{Read saved security metadata}
\usage{
.read_security_data(
  pkg_dir = ".",
  config_filename = "_beekeeper.yml",
  security_data_filename = .read_security_data_filename(pkg_dir, config_filename)
)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{security_data_filename}{(\code{character(1)} or \code{fs_path}) The path to the
saved security metadata file (relative to the package root).}
}
\value{
(\code{list}) Saved security metadata.
}
\description{
Read saved security metadata
}
\keyword{internal}
