% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-security.R
\name{.security_args_compile}
\alias{.security_args_compile}
\title{Compile security argument metadata}
\usage{
.security_args_compile(security_scheme_collection)
}
\arguments{
\item{security_scheme_collection}{(\code{list}) Collected security scheme
metadata.}
}
\value{
(\code{list}) Compiled security argument values.
}
\description{
Compile security argument metadata
}
\keyword{internal}
