% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{.security_scheme_collection_finalize}
\alias{.security_scheme_collection_finalize}
\title{Finalize collected security scheme data}
\usage{
.security_scheme_collection_finalize(security_scheme_collection)
}
\arguments{
\item{security_scheme_collection}{(\code{list}) Collected security scheme
metadata.}
}
\value{
(\code{list}) Finalized security metadata.
}
\description{
Finalize collected security scheme data
}
\keyword{internal}
