% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-security.R
\name{.security_scheme_description_fill}
\alias{.security_scheme_description_fill}
\title{Fill a missing security scheme description}
\usage{
.security_scheme_description_fill(
  security_scheme_description,
  security_scheme_type
)
}
\arguments{
\item{security_scheme_description}{(\code{character(1)}) A security scheme
description from the API definition.}

\item{security_scheme_type}{(\code{character(1)}) The security scheme type.}
}
\value{
(\code{character(1)}) A security scheme description.
}
\description{
Fill a missing security scheme description
}
\keyword{internal}
