% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-security.R
\name{.security_scheme_rotate}
\alias{.security_scheme_rotate}
\title{Convert one security scheme to a list}
\usage{
.security_scheme_rotate(
  security_scheme_name,
  security_scheme_details,
  security_scheme_description
)
}
\arguments{
\item{security_scheme_name}{(\code{character(1)}) The security scheme name.}

\item{security_scheme_details}{(\code{rapid::class_security_scheme_details}) The
security scheme details.}

\item{security_scheme_description}{(\code{character(1)}) A security scheme
description from the API definition.}
}
\value{
A \code{list} describing one security scheme.
}
\description{
Convert one security scheme to a list
}
\keyword{internal}
