% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{.security_schemes_collect}
\alias{.security_schemes_collect}
\title{Collect security scheme metadata}
\usage{
.security_schemes_collect(security_schemes)
}
\arguments{
\item{security_schemes}{(\code{rapid::class_security_schemes}) Security schemes
from the API definition.}
}
\value{
(\code{list}) Security scheme metadata.
}
\description{
Collect security scheme metadata
}
\keyword{internal}
