% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.setup_r}
\alias{.setup_r}
\title{Set up package directories and dependencies}
\usage{
.setup_r(pkg_dir)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{NULL}, invisibly) Called for setup side effects.
}
\description{
Set up package directories and dependencies
}
\keyword{internal}
