% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.stabilize_config}
\alias{.stabilize_config}
\title{Stabilize beekeeper config values}
\usage{
.stabilize_config(config)
}
\arguments{
\item{config}{(\code{list}) Package-generation configuration data.}
}
\value{
(\code{list}) A stabilized config list.
}
\description{
Stabilize beekeeper config values
}
\keyword{internal}
