% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.to_snake}
\alias{.to_snake}
\title{Convert text to snake case}
\usage{
.to_snake(x)
}
\arguments{
\item{x}{(\code{character}) Text to convert.}
}
\value{
(\code{character}) Snake-case text.
}
\description{
Convert text to snake case
}
\keyword{internal}
