% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.use_httptest2}
\alias{.use_httptest2}
\title{Ensure httptest2 is configured}
\usage{
.use_httptest2(pkg_dir)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{NULL}, invisibly) Called for setup side effects.
}
\description{
Ensure httptest2 is configured
}
\keyword{internal}
