% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.use_package}
\alias{.use_package}
\title{Add a package dependency to the DESCRIPTION file}
\usage{
.use_package(pkg, type = "Imports", pkg_dir = ".")
}
\arguments{
\item{pkg}{(\code{character(1)}) The package name.}

\item{type}{(\code{character(1)}) The dependency type, one of "Imports",
"Suggests", or "Depends".}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}, invisibly) The package name.
}
\description{
Add a package dependency to the DESCRIPTION file
}
\keyword{internal}
