% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.use_pkg_beekeeper}
\alias{.use_pkg_beekeeper}
\title{Add beekeeper to suggests}
\usage{
.use_pkg_beekeeper(pkg_dir = ".")
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}, invisibly) The package name.
}
\description{
Add beekeeper to suggests
}
\keyword{internal}
