% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.use_r_directory}
\alias{.use_r_directory}
\title{Ensure the R directory exists}
\usage{
.use_r_directory(pkg_dir)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{NULL}, invisibly) Called for setup side effects.
}
\description{
Ensure the R directory exists
}
\keyword{internal}
