% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.use_testthat}
\alias{.use_testthat}
\title{Ensure testthat is configured}
\usage{
.use_testthat(pkg_dir)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{NULL}, invisibly) Called for setup side effects.
}
\description{
Ensure testthat is configured
}
\keyword{internal}
