% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.use_tibblify}
\alias{.use_tibblify}
\title{Add tibblify to imports}
\usage{
.use_tibblify(pkg_dir = ".")
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}, invisibly) The package name.
}
\description{
Add tibblify to imports
}
\keyword{internal}
