% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{.without_security_file_path}
\alias{.without_security_file_path}
\title{Remove generated file paths from saved security metadata}
\usage{
.without_security_file_path(security_data)
}
\arguments{
\item{security_data}{(\code{list}) Generated security metadata.}
}
\value{
(\code{list}) Security metadata suitable for serialization.
}
\description{
Remove generated file paths from saved security metadata
}
\keyword{internal}
