% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beekeeper.R
\name{.write_config}
\alias{.write_config}
\title{Write the beekeeper config file}
\usage{
.write_config(
  api_definition,
  api_abbr,
  rapid_filename,
  config_filename,
  pkg_dir
)
}
\arguments{
\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{rapid_filename}{(\code{character(1)} or \code{fs_path}) The path to the R API
definition (rapid) file (relative to the package root).}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}) The written config file path.
}
\description{
Write the beekeeper config file
}
\keyword{internal}
