% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{.write_config_field}
\alias{.write_config_field}
\title{Update one config field}
\usage{
.write_config_field(field, value, config_filename, pkg_dir)
}
\arguments{
\item{field}{(\code{character(1)}) The config field to write.}

\item{value}{(\code{character(1)}) The value to write.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}) The written config filename.
}
\description{
Update one config field
}
\keyword{internal}
