% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beekeeper.R
\name{.write_rapid}
\alias{.write_rapid}
\title{Write the rapid definition file}
\usage{
.write_rapid(api_definition, rapid_file)
}
\arguments{
\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{rapid_file}{(\code{character(1)} or \code{fs_path}) The path to the R API
definition (rapid) file.}
}
\value{
(\code{character(1)}) The written file path.
}
\description{
Write the rapid definition file
}
\keyword{internal}
