% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{.write_security_data}
\alias{.write_security_data}
\title{Write saved security metadata and update the config}
\usage{
.write_security_data(
  security_data,
  security_data_filename,
  config_filename,
  pkg_dir
)
}
\arguments{
\item{security_data}{(\code{list}) Generated security metadata.}

\item{security_data_filename}{(\code{character(1)} or \code{fs_path}) The path to the
saved security metadata file (relative to the package root).}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}) The saved security metadata filename.
}
\description{
Write saved security metadata and update the config
}
\keyword{internal}
