% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg.R
\name{generate_pkg}
\alias{generate_pkg}
\title{Use a beekeeper config file to generate code}
\usage{
generate_pkg(
  api_abbr = NULL,
  api_definition = NULL,
  api_title = NULL,
  config_file = "_beekeeper.yml",
  pkg_dir = "."
)
}
\arguments{
\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{api_title}{(\code{character(1)}) The API title used in generated package
files.}

\item{config_file}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper yaml
config file.}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
A character vector of paths to files that were added or updated,
invisibly.
}
\description{
Creates or updates package files based on the information in a beekeeper
config file (generated by \code{\link[=use_beekeeper]{use_beekeeper()}} or manually). The files enforce
an opinionated framework for API packages.
}
