% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_agent.R
\name{generate_pkg_agent}
\alias{generate_pkg_agent}
\title{Create a user agent for the active package}
\usage{
generate_pkg_agent(path = ".")
}
\arguments{
\item{path}{The path to the DESCRIPTION file, or to a directory within a
package.}
}
\value{
A string with the name of the package and (if available) the first
URL associated with the package.
}
\description{
Create a user agent for the active package
}
