% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{generate_pkg_auth}
\alias{generate_pkg_auth}
\title{Generate authentication helpers}
\usage{
generate_pkg_auth(
  api_abbr = read_api_abbr(pkg_dir, config_filename),
  security_schemes = read_security_schemes(pkg_dir, read_rapid_filename(pkg_dir,
    config_filename)),
  save_security_data = TRUE,
  security_data_filename = "_beekeeper_security.yml",
  config_filename = "_beekeeper.yml",
  pkg_dir = "."
)
}
\arguments{
\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{security_schemes}{(\code{rapid::class_security_schemes}) Security schemes
from the API definition.}

\item{save_security_data}{(\code{logical(1)}) Whether to save generated security
metadata to a file and record that file in the beekeeper config.}

\item{security_data_filename}{(\code{character(1)} or \code{fs_path}) The path to the
saved security metadata file (relative to the package root).}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{list}) Generated security metadata.
}
\description{
Generate authentication helpers
}
