% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{oas_config}
\alias{oas_config}
\title{Configure a Package via an OpenAPI Spec}
\usage{
oas_config(oas, api_abbr)
}
\arguments{
\item{oas}{An openAPI spec, or the URL of such a spec.}

\item{api_abbr}{A short (2-5 letter) abbreviation for the API, for use in
function names and environment variables.}
}
\value{
The OAS spec, invisibly.
}
\description{
Configure a Package via an OpenAPI Spec
}
