% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-lengthless-default}
\alias{op-lengthless-default}
\alias{\%|0|\%}
\title{Default value for length 0}
\usage{
x \%|0|\% y
}
\arguments{
\item{x}{(\code{any}) Object to check.}

\item{y}{(\code{any}) Default value for \code{x}.}
}
\value{
If \code{!length(x)}, will return \code{x}; otherwise returns \code{y}.
}
\description{
Default value for length 0
}
\seealso{
Other empty operators:
\code{\link{op-no-char-default}},
\code{\link{op-null-continuation}},
\code{\link{op-null-default}}
}
\concept{empty operators}
\keyword{internal}
