% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-continuation}
\alias{op-null-continuation}
\alias{\%&&\%}
\title{Default value for non-\code{NULL}}
\usage{
x \%&&\% y
}
\arguments{
\item{x}{(\code{any}) The object to check or coerce.}

\item{y}{(\code{any}) The default value.}
}
\value{
If \code{x} is \code{NULL}, will return \code{x}; otherwise returns \code{y}.
}
\description{
Default value for non-\code{NULL}
}
\seealso{
Other empty operators:
\code{\link{op-lengthless-default}},
\code{\link{op-na-coalesce}},
\code{\link{op-no-char-default}},
\code{\link{op-null-default}}
}
\concept{empty operators}
\keyword{internal}
