% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Default value for \code{NULL}}
\usage{
x \%||\% y
}
\arguments{
\item{x}{(\code{any}) Object to check.}

\item{y}{(\code{any}) Default value for \code{x}.}
}
\value{
If \code{x} is \code{NULL}, will return \code{y}; otherwise returns \code{x}.
}
\description{
Default value for \code{NULL}
}
\seealso{
Other empty operators:
\code{\link{op-lengthless-default}},
\code{\link{op-no-char-default}},
\code{\link{op-null-continuation}}
}
\concept{empty operators}
\keyword{internal}
