% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_api_definition}
\alias{read_api_definition}
\title{Read an API definition file}
\usage{
read_api_definition(pkg_dir = ".", rapid_file = "_beekeeper_rapid.rds")
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{rapid_file}{(\code{character(1)} or \code{fs_path}) The path to the R API
definition (rapid) file.}
}
\value{
(\code{rapid::class_rapid}) The definition of the API.
}
\description{
Reads an RDS file (default name \verb{_beekeeper_rapid.rds}) with the API
definition stored as a \code{\link[rapid:class_rapid]{rapid::class_rapid()}} object. This file is generated
by \code{\link[=use_beekeeper]{use_beekeeper()}} based on an OpenAPI definition file.
}
