% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_api_title}
\alias{read_api_title}
\title{Read api_title config field}
\usage{
read_api_title(pkg_dir = ".", config_filename = "_beekeeper.yml")
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}
}
\value{
(\code{character(1)}) The \code{api_title} field from the beekeeper
config file.
}
\description{
Read the \code{api_title} field from a beekeeper config file.
}
\examples{
read_api_title(
  pkg_dir = fs::path_package("beekeeper"),
  config_filename = "example_config.yml"
)
}
\seealso{
Other config readers:
\code{\link[=read_api_abbr]{read_api_abbr()}},
\code{\link[=read_api_definition]{read_api_definition()}},
\code{\link[=read_config]{read_config()}},
\code{\link[=read_rapid_filename]{read_rapid_filename()}},
\code{\link[=read_security_data]{read_security_data()}},
\code{\link[=read_security_data_filename]{read_security_data_filename()}},
\code{\link[=read_security_schemes]{read_security_schemes()}}
}
\concept{config readers}
