% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{read_config}
\alias{read_config}
\title{Read a beekeeper config file}
\usage{
read_config(pkg_dir = ".", config_file = "_beekeeper.yml")
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_file}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper yaml
config file.}
}
\value{
A \code{list} of configuration information, with elements \code{api_title},
\code{api_abbr}, \code{api_version}, \code{rapid_file}, and \code{updated_on}.
}
\description{
Read a beekeeper config file
}
