% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_security_data}
\alias{read_security_data}
\title{Read saved security metadata}
\usage{
read_security_data(
  pkg_dir = ".",
  config_filename = "_beekeeper.yml",
  security_data_filename = read_security_data_filename(pkg_dir, config_filename)
)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{security_data_filename}{(\code{character(1)} or \code{fs_path}) The path to the
saved security metadata file (relative to the package root).}
}
\value{
(\code{list}) Saved security metadata.
}
\description{
Read prepared security metadata from a YAML file (default name
\verb{_beekeeper_security.yml}) generated via \code{\link[=generate_pkg_auth]{generate_pkg_auth()}}.
}
\examples{
read_security_data(pkg_dir = fs::path_package("beekeeper", "trello"))
}
\seealso{
Other config readers:
\code{\link[=read_api_abbr]{read_api_abbr()}},
\code{\link[=read_api_definition]{read_api_definition()}},
\code{\link[=read_api_title]{read_api_title()}},
\code{\link[=read_config]{read_config()}},
\code{\link[=read_rapid_filename]{read_rapid_filename()}},
\code{\link[=read_security_data_filename]{read_security_data_filename()}},
\code{\link[=read_security_schemes]{read_security_schemes()}}
}
\concept{config readers}
