% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_security_data_filename}
\alias{read_security_data_filename}
\title{Read saved security metadata filename}
\usage{
read_security_data_filename(pkg_dir = ".", config_filename = "_beekeeper.yml")
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}
}
\value{
(\code{character(1)}) The configured security metadata file path.
}
\description{
Read the \code{security_data_filename} field from a beekeeper config file.
}
\seealso{
Other config readers:
\code{\link[=read_api_abbr]{read_api_abbr()}},
\code{\link[=read_api_definition]{read_api_definition()}},
\code{\link[=read_api_title]{read_api_title()}},
\code{\link[=read_config]{read_config()}},
\code{\link[=read_rapid_filename]{read_rapid_filename()}},
\code{\link[=read_security_data]{read_security_data()}},
\code{\link[=read_security_schemes]{read_security_schemes()}}
}
\concept{config readers}
