% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_security_schemes}
\alias{read_security_schemes}
\title{Read security schemes from an API definition}
\usage{
read_security_schemes(
  pkg_dir = ".",
  rapid_filename = read_rapid_filename(pkg_dir)
)
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{rapid_filename}{(\code{character(1)} or \code{fs_path}) The path to the R API
definition (rapid) file (relative to the package root).}
}
\value{
(\code{rapid::class_security_schemes}) Security schemes from the API
definition.
}
\description{
Read the security schemes from an API definition stored as a
\code{\link[rapid:class_rapid]{rapid::class_rapid()}} object.
}
\seealso{
Other config readers:
\code{\link[=read_api_abbr]{read_api_abbr()}},
\code{\link[=read_api_definition]{read_api_definition()}},
\code{\link[=read_api_title]{read_api_title()}},
\code{\link[=read_config]{read_config()}},
\code{\link[=read_rapid_filename]{read_rapid_filename()}},
\code{\link[=read_security_data]{read_security_data()}},
\code{\link[=read_security_data_filename]{read_security_data_filename()}}
}
\concept{config readers}
