% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_if_not_exported.R
\name{skip_if_not_exported}
\alias{skip_if_not_exported}
\title{Skip a test if a function is not exported}
\usage{
skip_if_not_exported(fn_name, pkg = testthat::testing_package())
}
\arguments{
\item{fn_name}{(\code{character(1)}) The name of the function to check.}

\item{pkg}{(\code{character(1)}) The name of the package to check. Defaults to
the name of the package being tested via \code{\link[testthat:testing_package]{testthat::testing_package()}}. If
the package is \emph{not} the package currently being tested, it must be
installed on the testing system.}
}
\value{
(\code{NULL}, invisibly) Called for its side effect: skips the current
test if \code{fn_name} is not exported by \code{pkg}.
}
\description{
Skip tests when functions aren't exported by a package (yet). Useful for
auto-generated tests, so the tests are not active until the function has been
reviewed and exported.
}
\examples{
# In a generated test file, place this at the top of the test_that() block:
\dontrun{
test_that("my_fn() returns expected result", {
  skip_if_not_exported("my_fn")
  # ... rest of test
})
}
}
