% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beekeeper.R
\name{use_beekeeper}
\alias{use_beekeeper}
\title{Configure a package to use beekeeper}
\usage{
use_beekeeper(
  x,
  api_abbr,
  ...,
  pkg_dir = ".",
  config_filename = "_beekeeper.yml",
  rapid_filename = "_beekeeper_rapid.rds"
)
}
\arguments{
\item{x}{An object to use to define the configuration. It must be
translatable to a \code{\link[rapid:class_rapid]{rapid::class_rapid()}} object by \code{\link[rapid:as_rapid]{rapid::as_rapid()}}.
Usually this will be a url pointing to an OpenAPI document, or a list
generated by reading such a document.}

\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{...}{These dots are for future extensions and must be empty.}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{rapid_filename}{(\code{character(1)} or \code{fs_path}) The path to the R API
definition (rapid) file (relative to the package root).}
}
\value{
(\code{character(1)}, invisibly) The path to the configuration file. The
config file is written as a side effect of this function. The rapid object
is also written, and the path to that file (relative to \code{pkg_dir}) is saved
in the config file.
}
\description{
Create a configuration file for a package to use beekeeper. The configuration
file tracks information that will be used for generation of other functions,
and the timestamp when the configuration was last updated or used.
}
\examples{
\dontshow{if (rlang::is_installed("withr")) withAutoprint(\{ # examplesIf}
# Set up an empty package.
pkg_dir <- unclass(fs::path_norm(withr::local_tempdir()))
usethis::create_package(pkg_dir, open = FALSE, check_name = FALSE)
usethis::local_project(pkg_dir)

# Read an api definition. This could also simply be a URL to such a
# definition.
api_definition <- read_api_definition(fs::path_package("beekeeper", "guru"))

# Set up the package to use beekeeper with that definition.
use_beekeeper(api_definition, "guru")
read_config()
expect_equal(read_api_definition(), api_definition)

# Clean up.
withr::deferred_run()
\dontshow{\}) # examplesIf}
}
