% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bk_data.R, R/generate_pkg_auth.R,
%   R/generate_pkg_paths.R
\name{as_bk_data}
\alias{as_bk_data}
\alias{as_bk_data,any-method}
\alias{as_bk_data,rapid::security_schemes-method}
\alias{as_bk_data,rapid::security_scheme_details-method}
\alias{as_bk_data,rapid::api_key_security_scheme-method}
\alias{as_bk_data,rapid::paths-method}
\title{Prepare rapid objects for beekeeper}
\usage{
as_bk_data(x, ...)

## S7 method for class <any>
as_bk_data(x, ...)

## S7 method for class <rapid::security_schemes>
as_bk_data(x, ...)

## S7 method for class <rapid::security_scheme_details>
as_bk_data(x, ...)

## S7 method for class <rapid::api_key_security_scheme>
as_bk_data(x, ...)

## S7 method for class <rapid::paths>
as_bk_data(x, ...)
}
\arguments{
\item{x}{(\code{any}) The object to check or coerce.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
(\code{list}) A list.
}
\description{
Convert \code{rapid} objects to lists of properties to use in beekeeper templates.
}
\keyword{internal}
