% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-setup.R
\name{.assert_is_pkg}
\alias{.assert_is_pkg}
\title{Error if not in package}
\usage{
.assert_is_pkg(pkg_dir = usethis::proj_get())
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{NULL}, invisibly) Called for error side effect.
}
\description{
Error if not in package
}
\keyword{internal}
