% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-prepare.R
\name{.assert_is_pkg}
\alias{.assert_is_pkg}
\title{Error if not in package}
\usage{
.assert_is_pkg(base_path = usethis::proj_get())
}
\arguments{
\item{base_path}{The root URL of the current project.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Error if not in package
}
\keyword{internal}
