% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.flatten_params_df}
\alias{.flatten_params_df}
\title{Flatten parameter metadata}
\usage{
.flatten_params_df(params_df)
}
\arguments{
\item{params_df}{(\code{data.frame}) Parameter metadata in tabular form.}
}
\value{
(\code{data.frame}) Flattened parameter metadata.
}
\description{
Flatten parameter metadata
}
\keyword{internal}
