% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-pagination.R
\name{.generate_pagination}
\alias{.generate_pagination}
\title{Generate pagination files}
\usage{
.generate_pagination()
}
\value{
A \code{list} of pagination metadata used while generating path files.
}
\description{
Generate pagination files
}
\keyword{internal}
