% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.generate_paths}
\alias{.generate_paths}
\title{Generate files for API paths}
\usage{
.generate_paths(
  paths,
  api_abbr,
  security_data,
  pagination_data = list(),
  base_url
)
}
\arguments{
\item{paths}{(\code{rapid::class_paths}) API path definitions.}

\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{security_data}{(\code{list}) Generated security metadata.}

\item{pagination_data}{(\code{list}) Pagination metadata used while generating
path files.}

\item{base_url}{(\code{character(1)}) The base URL used in generated test helpers.}
}
\value{
(\code{character}) Generated file paths.
}
\description{
Generate files for API paths
}
\keyword{internal}
