% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.generate_paths_file}
\alias{.generate_paths_file}
\title{Generate one operation file}
\usage{
.generate_paths_file(
  path_operation,
  operation_id,
  api_abbr,
  security_data,
  use_prefix = FALSE,
  pkg_dir = "."
)
}
\arguments{
\item{path_operation}{(\code{list}) Template data for one operation.}

\item{operation_id}{(\code{character(1)}) The operation identifier used in
generated file names.}

\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{security_data}{(\code{list}) Generated security metadata.}

\item{use_prefix}{(\code{logical(1)}) Whether to include \code{api_abbr} as a prefix in
generated path function names. When \code{FALSE} (the default), functions are
named \code{{operation_id}} and \verb{req_\{operation_id\}}. When \code{TRUE}, functions are
named \verb{\{api_abbr\}_\{operation_id\}} and \verb{req_\{api_abbr\}_\{operation_id\}}.}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}) The generated file path.
}
\description{
Generate one operation file
}
\keyword{internal}
