% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.generate_paths_test_file}
\alias{.generate_paths_test_file}
\title{Generate one tag-level test file}
\usage{
.generate_paths_test_file(tag_operations, tag_name, api_abbr)
}
\arguments{
\item{tag_operations}{(\code{list}) Operations grouped under one tag.}

\item{tag_name}{(\code{character(1)}) The tag name.}

\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}
}
\value{
(\code{character(1)}) The generated test file path.
}
\description{
Generate one tag-level test file
}
\keyword{internal}
