% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{.generate_security_arg_help}
\alias{.generate_security_arg_help}
\title{Build security argument help entries}
\usage{
.generate_security_arg_help(security_scheme_collection, security_args)
}
\arguments{
\item{security_scheme_collection}{(\code{list}) Collected security scheme
metadata.}

\item{security_args}{(\code{character}) Security argument names to remove from
operation parameters (because they are defined once, separately).}
}
\value{
(\code{list}) Template-ready parameter help entries.
}
\description{
Build security argument help entries
}
\keyword{internal}
