% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-shared.R
\name{.generate_shared_params}
\alias{.generate_shared_params}
\title{Generate shared parameter docs}
\usage{
.generate_shared_params(security_data)
}
\arguments{
\item{security_data}{(\code{list}) Generated security metadata.}
}
\value{
(\code{character(1)}) The generated file path.
}
\description{
Generate shared parameter docs
}
\keyword{internal}
