% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_pkg}
\alias{.is_pkg}
\title{Check whether we're in a package}
\usage{
.is_pkg(base_path = usethis::proj_get())
}
\arguments{
\item{base_path}{The root URL of the current project.}
}
\value{
\code{TRUE} if the project is a package, \code{FALSE} if not.
}
\description{
Inspired by usethis:::is_package.
}
\keyword{internal}
