% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{.path_row_to_list}
\alias{.path_row_to_list}
\title{Convert one operation row to template data}
\usage{
.path_row_to_list(
  operation_id,
  endpoint,
  operation,
  operation_summary,
  operation_description,
  tags,
  parameters,
  responses,
  ...
)
}
\arguments{
\item{operation_id}{(\code{character(1)}) The operation identifier used in
generated file names.}

\item{endpoint}{(\code{character(1)}) The operation endpoint.}

\item{operation}{(\code{character(1)}) The HTTP method.}

\item{operation_summary}{(\code{character(1)}) The operation summary.}

\item{operation_description}{(\code{character(1)}) The operation description.}

\item{tags}{(\code{character} or \code{list}) Tags for all operations.}

\item{parameters}{(\code{data.frame}) Operation parameters.}

\item{...}{Additional columns, ignored.}
}
\value{
(\code{list}) One operation description.
}
\description{
Convert one operation row to template data
}
\keyword{internal}
