% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg-paths.R
\name{.remove_security_args}
\alias{.remove_security_args}
\title{Remove security parameters}
\usage{
.remove_security_args(params, security_args)
}
\arguments{
\item{params}{(\code{list}) Parameter metadata for a single operation.}

\item{security_args}{(\code{character}) Security argument names to remove from
operation parameters.}
}
\value{
A \code{list} of non-security parameters.
}
\description{
Remove security parameters
}
\keyword{internal}
