% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{.security_arg_description_clean}
\alias{.security_arg_description_clean}
\title{Format one security argument help entry}
\usage{
.security_arg_description_clean(security_arg_description, security_arg_name)
}
\arguments{
\item{security_arg_description}{(\code{character(1)}) The argument description.}

\item{security_arg_name}{(\code{character(1)}) The argument name.}
}
\value{
(\code{list}) A named list with \code{name} and \code{description}.
}
\description{
Format one security argument help entry
}
\keyword{internal}
