% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg.R
\name{generate_pkg}
\alias{generate_pkg}
\title{Use a beekeeper config file to generate code}
\usage{
generate_pkg(
  api_abbr = read_api_abbr(pkg_dir, config_filename),
  api_definition = read_api_definition(pkg_dir, read_rapid_filename(pkg_dir,
    config_filename)),
  api_title = read_api_title(pkg_dir, config_filename),
  use_prefix = FALSE,
  config_filename = "_beekeeper.yml",
  pkg_dir = "."
)
}
\arguments{
\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{api_title}{(\code{character(1)}) The API title used in generated package
files.}

\item{use_prefix}{(\code{logical(1)}) Whether to include \code{api_abbr} as a prefix
in generated path function names. When \code{FALSE} (the default), functions
are named \code{{{operation_id}}} and \verb{req_\{\{operation_id\}\}}. When \code{TRUE},
functions are named \verb{\{\{api_abbr\}\}_\{\{operation_id\}\}} and
\verb{req_\{\{api_abbr\}\}_\{\{operation_id\}\}}.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character}, invisibly) Paths to files that were added or updated.
}
\description{
Creates or updates package files based on the information in a beekeeper
config file (generated by \code{\link[=use_beekeeper]{use_beekeeper()}} or manually). The files enforce
an opinionated framework for API packages.
}
\examples{
\dontshow{if (rlang::is_installed("withr")) withAutoprint(\{ # examplesIf}
# Set up an empty package.
pkg_dir <- unclass(fs::path_norm(withr::local_tempdir()))
usethis::create_package(pkg_dir, open = FALSE, check_name = FALSE)
bk_files <- c("_beekeeper.yml", "_beekeeper_rapid.rds")
fs::file_copy(
  fs::path_package("beekeeper", "guru", bk_files),
  fs::path(pkg_dir, bk_files)
)
usethis::local_project(pkg_dir)

# Generate a package.
generate_pkg()
fs::dir_ls("R")
fs::dir_ls("tests/testthat")

# Clean up.
withr::deferred_run()
\dontshow{\}) # examplesIf}
}
\seealso{
Other package generation functions:
\code{\link[=generate_pkg_auth]{generate_pkg_auth()}},
\code{\link[=generate_pkg_paths]{generate_pkg_paths()}},
\code{\link[=generate_pkg_req_prepare]{generate_pkg_req_prepare()}},
\code{\link[=generate_pkg_shared_params]{generate_pkg_shared_params()}}
}
\concept{package generation functions}
