% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_main.R
\name{generate_pkg}
\alias{generate_pkg}
\title{Use a beekeeper config file to generate code}
\usage{
generate_pkg(
  config_file = "_beekeeper.yml",
  pkg_agent = generate_pkg_agent(config_file)
)
}
\arguments{
\item{config_file}{The path to a beekeeper yaml file.}

\item{pkg_agent}{A string to identify this package, for use in the
\code{user_agent} argument of \code{\link[nectar:call_api]{nectar::call_api()}}.}
}
\value{
A character vector of paths to files that were added or updated,
invisibly.
}
\description{
Creates or updates package files based on the information in a beekeeper
config file (generated by \code{\link[=use_beekeeper]{use_beekeeper()}} or manually). The files enforce
an opinionated framework for API packages.
}
