% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_auth.R
\name{generate_pkg_auth}
\alias{generate_pkg_auth}
\title{Generate authentication helpers}
\usage{
generate_pkg_auth(
  api_abbr = read_api_abbr(pkg_dir, config_filename),
  security_schemes = read_security_schemes(pkg_dir, read_rapid_filename(pkg_dir,
    config_filename)),
  save_security_data = TRUE,
  security_data_filename = "_beekeeper_security.yml",
  config_filename = "_beekeeper.yml",
  pkg_dir = "."
)
}
\arguments{
\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{security_schemes}{(\code{rapid::class_security_schemes}) Security schemes
from the API definition.}

\item{save_security_data}{(\code{logical(1)}) Whether to save generated security
metadata to a file and record that file in the beekeeper config.}

\item{security_data_filename}{(\code{character(1)} or \code{fs_path}) The path to the
saved security metadata file (relative to the package root).}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{list}) Generated security metadata. \code{R/020-auth.R} is generated as
a side effect. When \code{save_security_data} is \code{TRUE} (strongly recommended
when calling this function as a stand-alone), the file designated by
\code{security_data_filename}, and the \code{security_data_filename} field in the
file designated by \code{config_filename} are generated as side effects.
}
\description{
Generate the authentication helper file for a package under development from
the API security schemes stored in the OpenAPI definition. This supports
incremental package scaffolding when you want to review or customize auth
handling before generating the rest of the package files.
}
\seealso{
Other package generation functions:
\code{\link[=generate_pkg]{generate_pkg()}},
\code{\link[=generate_pkg_paths]{generate_pkg_paths()}},
\code{\link[=generate_pkg_req_prepare]{generate_pkg_req_prepare()}},
\code{\link[=generate_pkg_shared_params]{generate_pkg_shared_params()}}
}
\concept{package generation functions}
