% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_paths.R
\name{generate_pkg_paths}
\alias{generate_pkg_paths}
\title{Generate files for API paths}
\usage{
generate_pkg_paths(
  api_abbr = read_api_abbr(pkg_dir, config_filename),
  api_definition = read_api_definition(pkg_dir, read_rapid_filename(pkg_dir,
    config_filename)),
  security_data = read_security_data(pkg_dir, config_filename),
  config_filename = "_beekeeper.yml",
  pkg_dir = "."
)
}
\arguments{
\item{api_abbr}{(\code{character(1)}) A short (about 2-5 letter) abbreviation for
the API, for use in function names and environment variables.}

\item{api_definition}{(\code{rapid::class_rapid}) The API definition to generate
package code from.}

\item{security_data}{(\code{list}) Generated security metadata.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character}) Generated file paths. As a side effect, zero or more
\verb{R/paths-*.R} files are generated from the paths in the \code{api_definition},
as well as the associated test files as \verb{tests/testthat/test-paths-*.R}.
}
\description{
Generate operation functions and their tests from the API paths stored in the
OpenAPI definition. This supports incremental package scaffolding when you
want to review or customize endpoint wrappers separately from other package
components.
}
\seealso{
Other package generation functions:
\code{\link[=generate_pkg]{generate_pkg()}},
\code{\link[=generate_pkg_auth]{generate_pkg_auth()}},
\code{\link[=generate_pkg_req_prepare]{generate_pkg_req_prepare()}},
\code{\link[=generate_pkg_shared_params]{generate_pkg_shared_params()}}
}
\concept{package generation functions}
