% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pkg_shared_params.R
\name{generate_pkg_shared_params}
\alias{generate_pkg_shared_params}
\title{Generate shared parameter docs}
\usage{
generate_pkg_shared_params(
  security_data = read_security_data(pkg_dir, "_beekeeper.yml"),
  pkg_dir = "."
)
}
\arguments{
\item{security_data}{(\code{list}) Generated security metadata.}

\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}
}
\value{
(\code{character(1)}) The generated file path. As a side effect,
\code{R/000-shared.R} is generated.
}
\description{
Generate the shared roxygen parameter definitions used across scaffolded
functions. This supports incremental package scaffolding when you want to
review or customize the shared parameter topic independently while iterating
on the rest of the generated package files.
}
\examples{
\dontshow{if (rlang::is_installed("withr")) withAutoprint(\{ # examplesIf}
# Set up an empty package.
pkg_dir <- unclass(fs::path_norm(withr::local_tempdir()))
usethis::create_package(pkg_dir, open = FALSE, check_name = FALSE)
bk_files <- c("_beekeeper.yml", "_beekeeper_rapid.rds")
fs::file_copy(
  fs::path_package("beekeeper", "trello", bk_files),
  fs::path(pkg_dir, bk_files)
)
usethis::local_project(pkg_dir)

# Generate shared parameters.
generate_pkg_shared_params()
fs::dir_ls("R")

# Clean up.
withr::deferred_run()
\dontshow{\}) # examplesIf}
}
\seealso{
Other package generation functions:
\code{\link[=generate_pkg]{generate_pkg()}},
\code{\link[=generate_pkg_auth]{generate_pkg_auth()}},
\code{\link[=generate_pkg_paths]{generate_pkg_paths()}},
\code{\link[=generate_pkg_req_prepare]{generate_pkg_req_prepare()}}
}
\concept{package generation functions}
