% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-no-char-default}
\alias{op-no-char-default}
\alias{\%|a|\%}
\title{Default value for empty strings}
\usage{
x \%|a|\% y
}
\arguments{
\item{x}{(\code{any}) The object to check or coerce.}

\item{y}{(\code{character}) The default value.}
}
\value{
(\code{character}) If \code{!nzchar(x)}, returns \code{y}; otherwise returns \code{x}.
}
\description{
Default value for empty strings
}
\seealso{
Other empty operators:
\code{\link{op-lengthless-default}},
\code{\link{op-na-coalesce}},
\code{\link{op-null-continuation}},
\code{\link{op-null-default}}
}
\concept{empty operators}
\keyword{internal}
