% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Read a beekeeper config file}
\usage{
read_config(pkg_dir = ".", config_filename = "_beekeeper.yml")
}
\arguments{
\item{pkg_dir}{(\code{character(1)} or \code{fs_path}) The directory containing package
files.}

\item{config_filename}{(\code{character(1)} or \code{fs_path}) The path to a beekeeper
yaml config file (relative to the package root).}
}
\value{
(\code{list}) Configuration information, with elements \code{api_title},
\code{api_abbr}, \code{api_version}, \code{rapid_filename}, and \code{updated_on}, plus
\code{security_data_filename} when present.
}
\description{
Read a YAML file (default name \verb{_beekeeper.yml}) with configuration
information about the API and the package. This file is generated by
\code{\link[=use_beekeeper]{use_beekeeper()}}, or can be created manually.
}
\details{
The configuration information must contain these fields:
\itemize{
\item \code{api_title} (\code{character(1)}): The title of the API, used in documentation
and messages.
\item \code{api_abbr} (\code{character(1)}): An abbreviation for the API, used in function
names and other identifiers.
\item \code{api_version} (\code{character(1)}): The version of the API.
\item \code{rapid_filename} (\code{character(1)}): The name of the file (relative to the
package root) where the API definition is stored as an RDS file. By default,
this is \verb{_beekeeper_rapid.rds}, and is generated by \code{\link[=use_beekeeper]{use_beekeeper()}} based
on an OpenAPI definition file.
\item \code{updated_on} (\code{character(1)}): A timestamp of when the API definition was
last updated, in the format "YYYY-MM-DD HH:MM:SS" (UTC).
\item \code{security_data_filename} (\code{character(1)}): Optionally, the path (relative
to the package root) where generated security metadata is stored.
}
}
\seealso{
Other config readers:
\code{\link[=read_api_abbr]{read_api_abbr()}},
\code{\link[=read_api_definition]{read_api_definition()}},
\code{\link[=read_api_title]{read_api_title()}},
\code{\link[=read_rapid_filename]{read_rapid_filename()}},
\code{\link[=read_security_data]{read_security_data()}},
\code{\link[=read_security_data_filename]{read_security_data_filename()}},
\code{\link[=read_security_schemes]{read_security_schemes()}}
}
\concept{config readers}
