% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beekeeper.R
\name{use_beekeeper}
\alias{use_beekeeper}
\title{Configure a package to use beekeeper}
\usage{
use_beekeeper(
  x,
  api_abbr,
  ...,
  config_file = "_beekeeper.yml",
  rapid_file = "_beekeeper_rapid.rds"
)
}
\arguments{
\item{x}{An object to use to define the configuration. It must be
translatable to a \code{\link[rapid:class_rapid]{rapid::class_rapid()}} object by \code{\link[rapid:as_rapid]{rapid::as_rapid()}}.
Usually this will be a url pointing to an OpenAPI document, or a list
generated by reading such a document.}

\item{api_abbr}{A short (about 2-5 letter) abbreviation for the API, for use
in function names and environment variables.}

\item{...}{These dots are for future extensions and must be empty.}

\item{config_file}{The path to which the configuration should be written.}

\item{rapid_file}{The path to which the R API definition (rapid) object
should be written.}
}
\value{
The path to the configuration file, invisibly. The config file is
written as a side effect of this function. The rapid object is also
written, and the path to that file is saved in the config file.
}
\description{
Create a configuration file for a package to use beekeeper. The configuration
file tracks information that will be used for generation of other functions,
and the timestamp when the configuration was last updated or used.
}
